Allow embedding forms within other websites
Add embedded endpoint for page controller and allow
inserting submissions without CSFR as anonymous submissions
for public shares.

Added submenu entry for copying the embedding code to the clipboard
and added documentation on how to use the embedded view.

Switched to `vue-clipboard2` to allow copying to clipboard
from submenu entry (allows setting a container for the copy action).

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>